+++ app/views/oauth/applications/index.html.slim
@@ -1,32 +1,31 @@
-- title = "アプリケーション"
-- description = meta_description
-- keywords = meta_keywords
-- set_meta_tags(title: title, description: description, keywords: keywords, reverse: true)
+- title = t "head.title.oauth.applications.index"
+- set_meta_tags(title: title, og: { title: title }, reverse: true, prefix: false)
+- set_page_category "oauth_applications_index"
 
-= render "application/v2/navbar"
+= render "application/components/navbar"
 
-.ann-content.row.align-center
-  .small-12.columns
-    .ann-new-button.text-right
-      = link_to "新規登録", new_oauth_application_path, class: "button"
+.c-card.container.mt-1.mb-2.py-1
+  .text-xs-right.mb-1
+    = link_to new_oauth_application_path, class: "btn btn-primary" do
+      = icon "plus"
+      = t "messages.oauth.applications.new"
 
-    - if @applications.present?
-      table
-        thead
+  - if @applications.present?
+    table.table
+      thead
+        tr
+          th = t "activerecord.attributes.oauth_application.name"
+          th = t "activerecord.attributes.oauth_application.redirect_uri"
+          th = t "words.edit"
+          th = t "words.delete"
+      tbody
+        - @applications.each do |application|
           tr
-            th 名前
-            th コールバックURL
-            th 編集
-            th 削除
-        tbody
-          - @applications.each do |application|
-            tr
-              td = link_to application.name, oauth_application_path(application)
-              td = application.redirect_uri
-              td = link_to "編集", edit_oauth_application_path(application), class: "button warning"
-              td = link_to "削除", oauth_application_path(application), class: "button alert", data: { method: :delete, confirm: "削除しますか？" }
-    - else
-      .callout.secondary
-        p アプリケーションはありません
+            td = link_to application.name, oauth_application_path(application)
+            td = application.redirect_uri
+            td = link_to t("words.edit"), edit_oauth_application_path(application), class: "btn btn-secondary"
+            td = link_to t("words.delete"), oauth_application_path(application), class: "btn btn-danger", data: { method: :delete, confirm: t("messages._common.are_you_sure") }
+  - else
+    = render "application/components/empty", text: t("messages.oauth.applications.no_applications")
 
-= render "application/v2/components/footer"
+= render "application/components/footer"
